first potentially functioning mask creation code
