--- src/main/java/net/preibisch/bigstitcher/spark/fusion/WriteSuperBlock.java
@@ -1,252 +1,0 @@
-/*-
- * #%L
- * Spark-based parallel BigStitcher project.
- * %%
- * Copyright (C) 2021 - 2024 Developers.
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 2 of the
- * License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-2.0.html>.
- * #L%
- */
-package net.preibisch.bigstitcher.spark.fusion;
-
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-
-import org.apache.spark.api.java.function.VoidFunction;
-import org.janelia.saalfeldlab.n5.N5Writer;
-import org.janelia.saalfeldlab.n5.universe.N5Factory.StorageFormat;
-
-import mpicbg.spim.data.SpimData;
-import mpicbg.spim.data.registration.ViewRegistration;
-import mpicbg.spim.data.registration.ViewRegistrations;
-import mpicbg.spim.data.registration.ViewTransformAffine;
-import mpicbg.spim.data.sequence.ViewId;
-import net.imglib2.FinalInterval;
-import net.imglib2.Interval;
-import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.img.cell.CellGrid;
-import net.imglib2.realtransform.AffineTransform3D;
-import net.imglib2.type.NativeType;
-import net.imglib2.type.numeric.integer.UnsignedByteType;
-import net.imglib2.type.numeric.integer.UnsignedShortType;
-import net.imglib2.type.numeric.real.FloatType;
-import net.imglib2.util.Intervals;
-import net.imglib2.util.Util;
-import net.preibisch.bigstitcher.spark.blk.Fusion;
-import net.preibisch.bigstitcher.spark.blk.FusionFirstWins;
-import net.preibisch.bigstitcher.spark.blk.N5Helper;
-import net.preibisch.bigstitcher.spark.util.N5Util;
-import net.preibisch.bigstitcher.spark.util.Spark;
-import net.preibisch.bigstitcher.spark.util.ViewUtil;
-import net.preibisch.mvrecon.fiji.spimdata.SpimData2;
-
-public class WriteSuperBlock implements VoidFunction< long[][] >
-{
-	private static final long serialVersionUID = 4185504467908084954L;
-
-	private final URI xmlURI;
-
-	private final boolean preserveAnisotropy;
-
-	private final double anisotropyFactor;
-
-	private final long[] minBB;
-
-	private final URI n5PathURI;
-
-	private final String n5Dataset;
-
-	private final StorageFormat storageType;
-
-	private final int[][] serializedViewIds;
-
-	private final boolean uint8;
-
-	private final boolean uint16;
-
-	private final double minIntensity;
-
-	private final double range;
-
-	private final int[] blockSize;
-
-	private final boolean firstTileWins;
-
-	public WriteSuperBlock(
-			final URI xmlURI,
-			final boolean preserveAnisotropy,
-			final double anisotropyFactor,
-			final long[] minBB,
-			final URI n5PathURI,
-			final String n5Dataset,
-			final String bdvString,
-			final StorageFormat storageType,
-			final int[][] serializedViewIds,
-			final boolean uint8,
-			final boolean uint16,
-			final double minIntensity,
-			final double range,
-			final int[] blockSize,
-			final boolean firstTileWins )
-	{
-		this.xmlURI = xmlURI;
-		this.preserveAnisotropy = preserveAnisotropy;
-		this.anisotropyFactor = anisotropyFactor;
-		this.minBB = minBB;
-		this.n5PathURI = n5PathURI;
-		this.n5Dataset = n5Dataset;
-		this.storageType = storageType;
-		this.serializedViewIds = serializedViewIds;
-		this.uint8 = uint8;
-		this.uint16 = uint16;
-		this.minIntensity = minIntensity;
-		this.range = range;
-		this.blockSize = blockSize;
-		this.firstTileWins = firstTileWins;
-	}
-
-	@Override
-	public void call( final long[][] gridBlock ) throws Exception
-	{
-		final int n = blockSize.length;
-
-		// The min coordinates of the block that this job renders (in pixels)
-		final long[] superBlockOffset = new long[ n ];
-		Arrays.setAll( superBlockOffset, d -> gridBlock[ 0 ][ d ] + minBB[ d ] );
-
-		// The size of the block that this job renders (in pixels)
-		final long[] superBlockSize = gridBlock[ 1 ];
-
-		System.out.println( "Fusing block: offset=" + Util.printCoordinates( gridBlock[0] ) + ", dimension=" + Util.printCoordinates( gridBlock[1] ) );
-
-		// The min grid coordinate of the block that this job renders, in units of the output grid.
-		// Note, that the block that is rendered may cover multiple output grid cells.
-		final long[] outputGridOffset = gridBlock[ 2 ];
-
-
-
-
-		// --------------------------------------------------------
-		// initialization work that is happening in every job,
-		// independent of gridBlock parameters
-		// --------------------------------------------------------
-
-		// custom serialization
-		final SpimData2 dataLocal = Spark.getSparkJobSpimData2(xmlURI);
-		final List< ViewId > viewIds = Spark.deserializeViewIds( serializedViewIds );
-
-		// If requested, preserve the anisotropy of the data (such that
-		// output data has the same anisotropy as input data) by prepending
-		// an affine to each ViewRegistration
-		if ( preserveAnisotropy )
-		{
-			final AffineTransform3D aniso = new AffineTransform3D();
-			aniso.set(
-					1.0, 0.0, 0.0, 0.0,
-					0.0, 1.0, 0.0, 0.0,
-					0.0, 0.0, 1.0 / anisotropyFactor, 0.0 );
-			final ViewTransformAffine preserveAnisotropy = new ViewTransformAffine( "preserve anisotropy", aniso );
-
-			final ViewRegistrations registrations = dataLocal.getViewRegistrations();
-			for ( final ViewId viewId : viewIds )
-			{
-				final ViewRegistration vr = registrations.getViewRegistration( viewId );
-				vr.preconcatenateTransform( preserveAnisotropy );
-				vr.updateModel();
-			}
-		}
-
-
-		final long[] gridPos = new long[ n ];
-		final long[] fusedBlockMin = new long[ n ];
-		final long[] fusedBlockMax = new long[ n ];
-		final Interval fusedBlock = FinalInterval.wrap( fusedBlockMin, fusedBlockMax );
-
-		// pre-filter views that overlap the superBlock
-		Arrays.setAll( fusedBlockMin, d -> superBlockOffset[ d ] );
-		Arrays.setAll( fusedBlockMax, d -> superBlockOffset[ d ] + superBlockSize[ d ] - 1 );
-		final List< ViewId > overlappingViews = OverlappingViews.findOverlappingViews( dataLocal, viewIds, fusedBlock );
-
-		final N5Writer executorVolumeWriter = N5Util.createN5Writer(n5PathURI, storageType);//URITools.instantiateN5Writer( storageType, n5PathURI );//N5Util.createWriter( n5Path, storageType );
-		final ExecutorService prefetchExecutor = Executors.newCachedThreadPool(); //Executors.newFixedThreadPool( N_PREFETCH_THREADS );
-
-		final CellGrid blockGrid = new CellGrid( superBlockSize, blockSize );
-		final int numCells = ( int ) Intervals.numElements( blockGrid.getGridDimensions() );
-		for ( int gridIndex = 0; gridIndex < numCells; ++gridIndex )
-		{
-			blockGrid.getCellGridPositionFlat( gridIndex, gridPos );
-			blockGrid.getCellInterval( gridPos, fusedBlockMin, fusedBlockMax );
-
-			for ( int d = 0; d < n; ++d )
-			{
-				gridPos[ d ] += outputGridOffset[ d ];
-				fusedBlockMin[ d ] += superBlockOffset[ d ];
-				fusedBlockMax[ d ] += superBlockOffset[ d ];
-			}
-			// gridPos is now the grid coordinate in the N5 output
-
-			// determine which Cells and Views we need to compute the fused block
-			final OverlappingBlocks overlappingBlocks = OverlappingBlocks.find( dataLocal, overlappingViews, fusedBlock );
-
-			if ( overlappingBlocks.overlappingViews().isEmpty() )
-				continue;
-
-			try ( AutoCloseable prefetched = overlappingBlocks.prefetch( prefetchExecutor ) )
-			{
-				NativeType type;
-				if ( uint8 )
-					type = new UnsignedByteType();
-				else if ( uint16 )
-					type = new UnsignedShortType();
-				else
-					type = new FloatType();
-
-				final RandomAccessibleInterval< NativeType > source;
-
-				if ( firstTileWins )
-				{
-					source = FusionFirstWins.fuseVirtual(
-							dataLocal,
-							overlappingBlocks.overlappingViews(),
-							fusedBlock,
-							type,
-							minIntensity,
-							range );
-				}
-				else
-				{
-					source = Fusion.fuseVirtual(
-							dataLocal,
-							overlappingBlocks.overlappingViews(),
-							fusedBlock,
-							type,
-							minIntensity,
-							range );
-				}
-
-				N5Helper.saveBlock( source, executorVolumeWriter, n5Dataset, gridPos );
-			}
-		}
-		prefetchExecutor.shutdown();
-
-		// if it is not the shared HDF5 writer, then close
-		if ( N5Util.sharedHDF5Writer != executorVolumeWriter )
-			executorVolumeWriter.close();
-	}
-}
